Add Redis Sentinel support
